Complete CIS ROM in CPLD implementation.
It is good enough for autodetection of SRAM as memory at 100ns. So far the
RAM/disk switch just disables the CIS if set in "disk" position. Then the user
can use PrepCard to write his own settings into beginning of common memory
area. It needs further investigation if it makes sense to put "disk" CIS
into CPLD (the filesystem has to be created anyway).